+++ packages/compiler/src/core/diagnostic-creator.ts
@@ -1,3 +1,4 @@
+import { mutate } from "../utils/misc.js";
 import type { Program } from "./program.js";
 import type {
   Diagnostic,
@@ -50,12 +51,16 @@ export function createDiagnosticCreator<T extends { [code: string]: DiagnosticMe
 
     const messageStr = typeof message === "string" ? message : message((diagnostic as any).format);
 
-    return {
+    const result: Diagnostic = {
       code: libraryName ? `${libraryName}/${String(diagnostic.code)}` : diagnostic.code.toString(),
       severity: diagnosticDef.severity,
       message: messageStr,
       target: diagnostic.target,
     };
+    if (diagnostic.codefixes) {
+      mutate(result).codefixes = diagnostic.codefixes;
+    }
+    return result;
   }
 
   function reportDiagnostic<C extends keyof T, M extends keyof T[C] = "default">(