+++ packages/http/src/index.ts
@@ -1,6 +1,7 @@
 export { $lib } from "./lib.js";
 export { $linter } from "./linter.js";
 
+export * from "./auth.js";
 export * from "./content-types.js";
 export * from "./decorators.js";
 export * from "./metadata.js";