+++ packages/internal-build-utils/CHANGELOG.md
@@ -1,12 +1,22 @@
 # Change Log - @typespec/internal-build-utils
 
+## 0.54.0
+
+### Bug Fixes
+
+- [#2914](https://github.com/microsoft/typespec/pull/2914) Bumping PR version will also update the dependencies to be an open range
+
+### Bump dependencies
+
+- [#2900](https://github.com/microsoft/typespec/pull/2900) Update dependencies
+
+
 ## 0.53.0
 
 ### Minor Changes
 
 - 4329c78: Change prerelease version bump logic to work with pnpm instead of rush
 
-This log was last generated on Wed, 24 Jan 2024 05:46:53 GMT and should not be manually modified.
 
 ## 0.52.0